requirements.txt
db-sqlite3==0.0.1
fastapi==0.101.1
nanoid
openpyxl
pandas==2.0.3
pydantic==2.3.0
pyserial
pytz==2023.3
requests
RPi.GPIO
scikit-learn
uvicorn==0.22.0